+++ src/layouts/MainNavLayout.tsx
@@ -39,6 +39,7 @@ import {
   ONLY_DEV_DESIGN_SYSTEM_ROUTE,
   ONLY_DEV_DESIGN_SYSTEM_TAB_ROUTE,
   ORGANIZATION_INFORMATIONS_ROUTE,
+  PAYMENTS_ROUTE,
   PLAN_DETAILS_ROUTE,
   PLAN_SUBSCRIPTION_DETAILS_ROUTE,
   PLANS_ROUTE,
@@ -358,6 +359,13 @@ const MainNavLayout = () => {
                     canBeClickedOnActive: true,
                     match: [INVOICES_ROUTE, INVOICES_TAB_ROUTE],
                     hidden: !hasPermissions(['invoicesView', 'creditNotesView']),
+                  },
+                  {
+                    title: translate('text_6672ebb8b1b50be550eccbed'),
+                    icon: 'coin-dollar',
+                    link: PAYMENTS_ROUTE,
+                    canBeClickedOnActive: true,
+                    hidden: !hasPermissions(['paymentsView']),
                   },
                 ]}
               />
