+++ src/pages/InvoicesPage.tsx
@@ -1,6 +1,6 @@
 import { gql } from '@apollo/client'
 import { useMemo, useRef } from 'react'
-import { generatePath, useNavigate, useParams, useSearchParams } from 'react-router-dom'
+import { generatePath, useParams, useSearchParams } from 'react-router-dom'
 
 import CreditNotesTable from '~/components/creditNote/CreditNotesTable'
 import { Button, NavigationTab, Typography } from '~/components/designSystem'
@@ -19,14 +19,13 @@ import {
   FinalizeInvoiceDialogRef,
 } from '~/components/invoices/FinalizeInvoiceDialog'
 import InvoicesList from '~/components/invoices/InvoicesList'
-import { PaymentsList } from '~/components/invoices/PaymentsList'
 import { VoidInvoiceDialog, VoidInvoiceDialogRef } from '~/components/invoices/VoidInvoiceDialog'
 import { PremiumWarningDialog, PremiumWarningDialogRef } from '~/components/PremiumWarningDialog'
 import { SearchInput } from '~/components/SearchInput'
 import { addToast, hasDefinedGQLError } from '~/core/apolloClient'
 import { INVOICE_LIST_FILTER_PREFIX } from '~/core/constants/filters'
 import { InvoiceListTabEnum } from '~/core/constants/tabsOptions'
-import { CREATE_PAYMENT_ROUTE, INVOICES_ROUTE, INVOICES_TAB_ROUTE } from '~/core/router'
+import { INVOICES_ROUTE, INVOICES_TAB_ROUTE } from '~/core/router'
 import { serializeAmount } from '~/core/serializers/serializeAmount'
 import {
   CreditNoteExportTypeEnum,
@@ -36,16 +35,13 @@ import {
   InvoiceExportTypeEnum,
   InvoiceListItemFragmentDoc,
   LagoApiError,
-  PaymentForPaymentsListFragmentDoc,
   useCreateCreditNotesDataExportMutation,
   useCreateInvoicesDataExportMutation,
   useGetCreditNotesListLazyQuery,
   useGetInvoicesListLazyQuery,
-  useGetPaymentListLazyQuery,
   useRetryAllInvoicePaymentsMutation,
 } from '~/generated/graphql'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
-import { useCurrentUser } from '~/hooks/useCurrentUser'
 import { useDebouncedSearch } from '~/hooks/useDebouncedSearch'
 import { useOrganizationInfos } from '~/hooks/useOrganizationInfos'
 import { usePermissions } from '~/hooks/usePermissions'
@@ -98,24 +94,6 @@ gql`
       collection {
         id
         ...InvoiceListItem
-      }
-    }
-  }
-
-  query getPaymentList($invoiceId: ID, $externalCustomerId: ID, $limit: Int, $page: Int) {
-    payments(
-      invoiceId: $invoiceId
-      externalCustomerId: $externalCustomerId
-      limit: $limit
-      page: $page
-    ) {
-      metadata {
-        currentPage
-        totalPages
-        totalCount
-      }
-      collection {
-        ...PaymentForPaymentsList
       }
     }
   }
@@ -181,7 +159,6 @@ gql`
   ${InvoiceListItemFragmentDoc}
   ${CreditNoteTableItemFragmentDoc}
   ${CreditNotesForTableFragmentDoc}
-  ${PaymentForPaymentsListFragmentDoc}
 `
 
 // TODO: This is a temporary workaround
@@ -214,8 +191,6 @@ const InvoicesPage = () => {
   const { translate } = useInternationalization()
   const { hasPermissions } = usePermissions()
   const { organization } = useOrganizationInfos()
-  const navigate = useNavigate()
-  const { isPremium } = useCurrentUser()
   const [searchParams] = useSearchParams()
   const amountCurrency = organization?.defaultCurrency
   const { tab = InvoiceListTabEnum.invoices } = useParams<{ tab?: InvoiceListTabEnum }>()
@@ -255,24 +230,6 @@ const InvoicesPage = () => {
   })
 
   const [
-    getPayments,
-    {
-      data: dataPayments,
-      loading: loadingPayments,
-      error: errorPayments,
-      fetchMore: fetchMorePayments,
-      variables: variablePayments,
-    },
-  ] = useGetPaymentListLazyQuery({
-    notifyOnNetworkStatusChange: true,
-    fetchPolicy: 'network-only',
-    nextFetchPolicy: 'network-only',
-    variables: {
-      limit: 20,
-    },
-  })
-
-  const [
     getCreditNotes,
     {
       data: dataCreditNotes,
@@ -295,8 +252,6 @@ const InvoicesPage = () => {
     useDebouncedSearch(getInvoices, loadingInvoices)
   const { debouncedSearch: creditNoteDebounceSearch, isLoading: creditNoteIsLoading } =
     useDebouncedSearch(getCreditNotes, loadingCreditNotes)
-  const { debouncedSearch: paymentsDebounceSearch, isLoading: paymentsIsLoading } =
-    useDebouncedSearch(getPayments, loadingPayments)
 
   const [retryAll] = useRetryAllInvoicePaymentsMutation({
     context: { silentErrorCodes: [LagoApiError.PaymentProcessorIsCurrentlyHandlingPayment] },
@@ -394,27 +349,6 @@ const InvoicesPage = () => {
               </Button>
             </>
           )}
-          {tab === InvoiceListTabEnum.payments && (
-            <>
-              <SearchInput
-                onChange={paymentsDebounceSearch}
-                placeholder={translate('text_17370296250897aidak5kjcg')}
-              />
-              <Button
-                variant="primary"
-                onClick={() => {
-                  if (isPremium) {
-                    navigate(CREATE_PAYMENT_ROUTE)
-                  } else {
-                    premiumWarningDialogRef.current?.openDialog()
-                  }
-                }}
-                endIcon={isPremium ? undefined : 'sparkles'}
-              >
-                {translate('text_1737471851634wpeojigr27w')}
-              </Button>
-            </>
-          )}
           {tab === InvoiceListTabEnum.creditNotes && (
             <>
               <SearchInput
@@ -475,23 +409,7 @@ const InvoicesPage = () => {
             ),
             hidden: !hasPermissions(['invoicesView']),
           },
-          {
-            title: translate('text_6672ebb8b1b50be550eccbed'),
-            link: generatePath(INVOICES_TAB_ROUTE, {
-              tab: InvoiceListTabEnum.payments,
-            }),
-            match: [generatePath(INVOICES_TAB_ROUTE, { tab: InvoiceListTabEnum.payments })],
-            component: (
-              <PaymentsList
-                error={errorPayments}
-                fetchMore={fetchMorePayments}
-                payments={dataPayments?.payments?.collection}
-                isLoading={paymentsIsLoading}
-                metadata={dataPayments?.payments?.metadata}
-                variables={variablePayments}
-              />
-            ),
-          },
+
           {
             title: translate('text_636bdef6565341dcb9cfb125'),
             link: generatePath(INVOICES_TAB_ROUTE, {
